requirements.txt
PyPDF2
scikit-learn >= 1.4.0
sentence-transformers
faiss-cpu
numpy
groq